Updates to Inventory
The add and edit product is now working as well as the drop downs for the location and type. I also modified the Inventory page by adding a button to add new products and removed the details button in to make better use of the screen by adding hoverable/clickable table rows. I also removed the damaged printout in favor of a checkmark for damaged items.

Back end I had to change some of the table names to make the dropdowns function.