Password Verification
I've made the changes to the password verification. By default verify is hidden. If the password has changed, the verify will now display with a red icon when not matching, if matching it turns green. The verify is always visible on add new user. I also added some logic to allow an edit as long as the original password has not been modified. This sops you from entering the verify every time. I also noticed that some of the fields were carrying over after a modal was closed so I added that they be cleared on the close.