Access Level Authentication
Moved the login methods into a new file to store the logged-in user's data in a sessionStorage. Fixed the routing of the application so that only login page can be viewed unless a user is logged-in. Added the logic to prevent non-admin users from adding, editing, or assigning products and from viewing other employees data.